+++ fastapi/openapi/models.py
@@ -2,6 +2,7 @@ from enum import Enum
 from typing import Any, Callable, Dict, Iterable, List, Optional, Union
 
 from fastapi.logger import logger
+from fastapi.security.models import APIKey, HTTPBase, HTTPBearer, OAuth2, OpenIdConnect
 from pydantic import AnyUrl, BaseModel, Field
 
 try:
@@ -274,88 +275,6 @@ class PathItem(BaseModel):
 
     class Config:
         extra = "allow"
-
-
-class SecuritySchemeType(Enum):
-    apiKey = "apiKey"
-    http = "http"
-    oauth2 = "oauth2"
-    openIdConnect = "openIdConnect"
-
-
-class SecurityBase(BaseModel):
-    type_: SecuritySchemeType = Field(alias="type")
-    description: Optional[str] = None
-
-    class Config:
-        extra = "allow"
-
-
-class APIKeyIn(Enum):
-    query = "query"
-    header = "header"
-    cookie = "cookie"
-
-
-class APIKey(SecurityBase):
-    type_ = Field(SecuritySchemeType.apiKey, alias="type")
-    in_: APIKeyIn = Field(alias="in")
-    name: str
-
-
-class HTTPBase(SecurityBase):
-    type_ = Field(SecuritySchemeType.http, alias="type")
-    scheme: str
-
-
-class HTTPBearer(HTTPBase):
-    scheme = "bearer"
-    bearerFormat: Optional[str] = None
-
-
-class OAuthFlow(BaseModel):
-    refreshUrl: Optional[str] = None
-    scopes: Dict[str, str] = {}
-
-    class Config:
-        extra = "allow"
-
-
-class OAuthFlowImplicit(OAuthFlow):
-    authorizationUrl: str
-
-
-class OAuthFlowPassword(OAuthFlow):
-    tokenUrl: str
-
-
-class OAuthFlowClientCredentials(OAuthFlow):
-    tokenUrl: str
-
-
-class OAuthFlowAuthorizationCode(OAuthFlow):
-    authorizationUrl: str
-    tokenUrl: str
-
-
-class OAuthFlows(BaseModel):
-    implicit: Optional[OAuthFlowImplicit] = None
-    password: Optional[OAuthFlowPassword] = None
-    clientCredentials: Optional[OAuthFlowClientCredentials] = None
-    authorizationCode: Optional[OAuthFlowAuthorizationCode] = None
-
-    class Config:
-        extra = "allow"
-
-
-class OAuth2(SecurityBase):
-    type_ = Field(SecuritySchemeType.oauth2, alias="type")
-    flows: OAuthFlows
-
-
-class OpenIdConnect(SecurityBase):
-    type_ = Field(SecuritySchemeType.openIdConnect, alias="type")
-    openIdConnectUrl: str
 
 
 SecurityScheme = Union[APIKey, HTTPBase, OAuth2, OpenIdConnect, HTTPBearer]
