+++ fastapi/security/api_key.py
@@ -1,7 +1,7 @@
 from typing import Optional
 
-from fastapi.openapi.models import APIKey, APIKeyIn
 from fastapi.security.base import SecurityBase
+from fastapi.security.models import APIKey, APIKeyIn
 from starlette.exceptions import HTTPException
 from starlette.requests import Request
 from starlette.status import HTTP_403_FORBIDDEN
