+++ fastapi/security/http.py
@@ -3,9 +3,9 @@ from base64 import b64decode
 from typing import Optional
 
 from fastapi.exceptions import HTTPException
-from fastapi.openapi.models import HTTPBase as HTTPBaseModel
-from fastapi.openapi.models import HTTPBearer as HTTPBearerModel
 from fastapi.security.base import SecurityBase
+from fastapi.security.models import HTTPBase as HTTPBaseModel
+from fastapi.security.models import HTTPBearer as HTTPBearerModel
 from fastapi.security.utils import get_authorization_scheme_param
 from pydantic import BaseModel
 from starlette.requests import Request
